Refactor `Taskt` into `Todo`
