+++ src/lib/ui/TodoItems.svelte
@@ -2,7 +2,7 @@
 	import type {Readable} from 'svelte/store';
 
 	import type {Entity} from '$lib/vocab/entity/entity';
-	import TaskItem from '$lib/ui/TaskItem.svelte';
+	import TodoItem from '$lib/ui/TodoItem.svelte';
 
 	export let entities: Readable<Readable<Entity>[]>;
 </script>
@@ -10,6 +10,6 @@
 <!-- TODO possibly remove the `ul` wrapper and change the `li`s to `div`s -->
 <ul>
 	{#each $entities as entity (entity)}
-		<TaskItem {entity} />
+		<TodoItem {entity} />
 	{/each}
 </ul>
