+++ src/lib/ui/view/components.ts
@@ -8,7 +8,7 @@ import Notes from '$lib/ui/view/Notes.svelte';
 import Iframe from '$lib/ui/view/Iframe.svelte';
 import Voice from '$lib/ui/view/Voice.svelte';
 import EntityExplorer from '$lib/ui/view/EntityExplorer.svelte';
-import Task from '$lib/ui/view/Task.svelte';
+import Todo from '$lib/ui/view/Todo.svelte';
 
 export const viewComponents: Record<string, typeof SvelteComponent> = {
 	Home,
@@ -19,5 +19,5 @@ export const viewComponents: Record<string, typeof SvelteComponent> = {
 	Voice,
 	Iframe,
 	EntityExplorer,
-	Task,
+	Todo,
 };
